chore :: 훅 분리
